requirements.txt
requests
python-dotenv
openpyxl
python-docx
